ARM: dts: qcom-msm8974: add modem node
This patch adds a node for the modem. It is identical to the one provided
the in the devicetree documentation, but with a fixed compatible and added
xo clock.

The reserved regions are modified to match those on hammerhead firmware.
The total reserved region doesn't change, only the start of mba_region.

Also adds the required qcom,rmtfs-mem compatible to the rmtfs memory.

Signed-off-by: Jonathan Marek <jonathan@marek.ca>