lib: optimize cpumask_local_spread()
In a multi-processor and NUMA system, I/O device may have many numa nodes
belonging to multiple cpus.  When we get a local numa, it is better to
find the node closest to the local numa node, instead of choosing any
online cpu immediately.

For the current code, it only considers the local NUMA node and it doesn't
compute the distances between different NUMA nodes for the non-local NUMA
nodes.  Let's optimize it and find the nearest node through NUMA distance.
The performance will be better if it returns the nearest node than a
random node.

When Parameter Server workload is tested using NIC device on Huawei
Kunpeng 920 SoC:
Without the patch, the performance is 22W QPS;
Added this patch, the performance become better and it is 26W QPS.

Link: http://lkml.kernel.org/r/1572863268-28585-1-git-send-email-zhangshaokun@hisilicon.com
Signed-off-by: yuqi jin <jinyuqi@huawei.com>
Signed-off-by: Shaokun Zhang <zhangshaokun@hisilicon.com>
Cc: Mike Rapoport <rppt@linux.ibm.com>
Cc: Paul Burton <paul.burton@mips.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Anshuman Khandual <anshuman.khandual@arm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>