drm/msm: revert async commit support
This patch reverts the following three commits that introduced async
commit support in the MSM driver:

    d934a712c5e6 ("drm/msm: add atomic traces")
    2d99ced787e3 ("drm/msm: async commit support")
    194fc68d7a49 ("drm/msm/dpu: async commit support")

Since this was introduced, the command mode panel on the Nexus 5 began
reporting the following errors:

    msm fd900000.mdss: pp done time out, lm=0

modetest from the libdrm tests reports that the page flips drop with
this issue present:

    $ modetest -v -s 32:1080x1920
    trying to open device 'msm'...done
    setting mode 1080x1920-71.71Hz@XR24 on connectors 32, crtc 50
    freq: 13.50Hz
    freq: 13.51Hz
    freq: 13.51Hz

This revert fixes the 'pp done time out' errors and gets the panel
running at the expected refresh rate again:

    $ modetest -v -s 32:1080x19
    trying to open device 'msm'...done
    setting mode 1080x1920-71.71Hz@XR24 on connectors 32, crtc 50
    failed to set gamma: Function not implemented
    freq: 59.40Hz
    freq: 59.69Hz
    freq: 59.69Hz

This isn't an ideal long-term solution. One possible fix is to enable
the autorefresh feature as described on this patch:

https://lore.kernel.org/lkml/20191230020053.26016-2-masneyb@onstation.org/

However the patch is not fully working as expected yet. More background
discussion is available at:

https://lore.kernel.org/dri-devel/20191105000129.GA6536@onstation.org/

Signed-off-by: Brian Masney <masneyb@onstation.org>