FROMLIST v3 ARM: dts: qcom: msm8974: Split out common part of samsung-klte
Samsung Galaxy S5 has many variants. Variants that support LTE use klte*
as their codename. Currently, the only supported one is the one without
any suffix, namely, klte. It is known that other klte* variants have
only minor differences compared to klte and can mostly work with the
klte DTB.

Split the common part into a common DTSI so that it can be imported in
the DTS of klte and other klte* variants.

Signed-off-by: Rong Zhang <i@rong.moe>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>